fond_10B3C.c attempts up to FUN_8013613c

+++ include/fond_10B3C.h
@@ -13,7 +13,7 @@ void PS1_LoadFondDataAndPalettes(void);
 void PS1_LoadFond(void);
 void FUN_80135ab0(s16 param_1, s32 param_2);
 void FUN_80135d5c(s32 param_1, s32 param_2, s32 param_3, s16 param_4);
-void FUN_8013613c(u16 param_1, u32 param_2, u16 param_3, u32 param_4);
+void FUN_8013613c(u8 param_1, u32 param_2, u8 param_3, u32 param_4);
 void FUN_80136340(s32 param_1, u32 param_2);
 void FUN_801366ac(void);
 void PS1_DisplayFondSprites(void);
